FOI-15: Prevent direct access to mid-journey pages
Background to this approach:

This simply ensures all users begin their journey at the service
entry point. From that point it relies on the application routing.

It might be that a future approach enforces a specific path through
the user journey, but my sense is that could get very difficult to
reason about very quickly. One thing I'm considering here is how
easy it will be for future developers (including myself) to understand
what's going on.

A note on "state machines":

I'd initially be considering a some way of potentially using a state
machine approach because I've seen this used elsewhere[^1] but I'm
currently unconvinced partly because these can get huge and partly
because I'm feeling like the application is always stateless at the
point a HTTP request is received. I'll keep thinking about this though.

What this commit does:

Adds session key protection to prevent access to mid-journey pages
without a key which is on the session when visiting the main.index
route.

1. routes.py is updated to set entered_through_index_page on the
   session
2. requires_session_key.py adds a function to the before_request
   in Flask which checks for the key on all but exempt_routes and
   a specific path we're using to check for 308 redirects
3. __init__.py is updated to call requires_session_key(app)

In addition to the functionality:

1. test_routes.py is updated to introduce two tests for
   requires_session_key
2. prevent-direct-mid-journey-access.spec.ts is introduced to add
   Playwright tests for the route protection mechanism.

[^1]: https://github.com/govuk-one-login/authentication-frontend/blob/main/src/components/common/state-machine/state-machine.ts