Refactor and improve comments for state machine
I've simply changed the convention of using '_statement' for non-form
pages to be '_page', because it occurred to me this is a little less
jargony while I was reviewing the comments which describe how the
state machine works.

Having done this, I also had to update the routes and tests - but it's
just a renaming.

The one additional change is updating the app title to reflect the latest
service name