FOI-57: Visually hide the legend
The Design System doesn't currently support the intended design, so -
having checked with the Lead Frontend Developer - I've introduced a
style to visually hide the legend. My sense is this will be a good
fix because I've received feedback during a11y audits previously
that a legend that repeats the h1 is an annoyance to users of
assistive technology (although that was one tester in one report).