FOI-63: Introduce with_form_prefilled_from_session decorator
This decorator is part of the work to introduce back links.

Because clicking a 'back' link should show the user's previous form selection, we need
to store the user's selections (while handling 'POST' requests) and present these (in
response to the 'GET' requests).

I was keen to avoid too much duplication here and wanted to keep our route functions
easy to read, so I opted to extract the instantiation of all forms to this decorator

I've also introduced tests:

1. test_get_prefills_form_from_session makes a GET request. It puts some data (foo
   and csrf_token) in the session, then checks if the form is prefilled with foo
   (but not csrf_token) when the page is loaded.
2. test_post_does_not_prefill_form makes a POST request. It puts foo in the session,
   then checks that the form is not prefilled with foo on POST requests.