FOI-70: Add "We may hold this record page"
This introduces the template, route function, content in YAML and route (in
constants.py).

As part of this change, I've introduced a new parse_markdown_links Jinja2
filter that returns a HTML link when the input string contains a markdown link.
Tests for this new filter have been added to test_template_filters.py