Add a .git-blame-ignore-revs file
I've added this file to hold the commit hashes for those commits we want to
exclude from 'git blame'. Having created the file you simply set your Git
config to recognise with a command like this:

git config blame.ignoreRevsFile .git-blame-ignore-revs

You can read about this in the Git blame docs[^1] and
in the Git config docs[^2]

Why am I doing this now? Because I'm about to run the Black formatter against
this codebase and have seen it's going to make a load of changes.

[^1]: https://git-scm.com/docs/git-blame#Documentation/git-blame.txt---ignore-revrev
[^2]: https://git-scm.com/docs/git-config#Documentation/git-config.txt-blameignoreRevsFile