FOI-71: Disambiguate (split) "We don't have" pages
Prior to this commit, one template and route was being used
for both:

1. We don't have records for this service branch
2. We don't have records for this rank

Looking at version 1.2 of the user flow, there are additional
variants and I feel there's a strong likelihood these pages will
need content specific to the relevant journey. I've therefore
decided to split them out so that we have different templates and
tests.