Rename state machine event for consistency
While working on FOI-72, I noticed that all but one Events in
state_machine.py are described as being "continue_from...". This makes
sense because events can have conditions and guards, which we use to
determine where the user will go. My sense is that there's value in
retaining this consistency - just to encourage clear thinking really.