FOI-74: Add routing from upload proof of death
Updates state machine to allow user progress from the 'Upload
a proof of death' to 'About the service person'. This includes
updates to the Pyton and Playwright tests.

I've also included a comment in the route function because there
is a user experience question to be answered, and I think it
might easily be overlooked during implementation.