Add function to save form data to session
While reviewing our route functions, it occurred to me that the approach
of each function saving the relevant form fields is a bit brittle.

In this commit, I've introduced the save_submitted_form_fields_to_session
which can do this for all route methods. I think this is a much better
approach because:

1. it will automatically pick up any additional fields added to forms
2. it consolidates the logic into a single function (which has comprehensive
   unit tests in test_save_submitted_form_fields_to_session.py)
3. it makes the route functions themselves easier to read