FOI-114: Save catalogue reference to session
This commit introduces the `save_catalogue_reference_to_session`
function. This function saves the "catalogue_reference" query
parameter to the session if it is present after it has been
sanitized.

While this isn't strictly "form" data, it seemed most sensible
for it to be saved alongside other fields obtained from the user
via form entries.