Add additional content tests
While reviewing the content yesterday I spotted a value in content.yaml
that was no longer needed. This commit introduces a couple of additional
tests to ensure the YAML stays up to date:

1. test_content_used_in_template() checks that every content key
   we have in the YAML is actually used in a template (it skips fields)
2. test_for_fields_passed_to_get_field_content() checks that all form
   field content is actually referenced in a form

My hope is that these test help ensure the content.yaml contains only
the content which is actually used in the app.