Add decorator to manage back links
The latest designs have pages that can be reached from several parts of the
journey (an example is the "Are you sure you want to cancel?" page). Because
we are using Post -> Redirect -> Get, managing the destination for the 'Back'
links on these pages is a bit of a challenge.

I'd initially considered whether we might be able to achieve this in some way
using the state machine, but that is also limited to the Request / Response
lifecycle.

The only way to achieve this server-side (and I'd rather not rely on client-
side JavaScript for obvious reasons[^1]) is to introduce a mechanism to save the
route to the session. In this way, when a user reaches a 'gateway' page (by
which I mean a page that can result in the user hitting, say, "Are you sure you
want to cancel?" in their next step, we can write the route to the session.

I think this is a pretty good solution, but we'll have to see if it works out.

[^1]: https://design-system.service.gov.uk/components/back-link/