Update back links for officer record routes and update tests
Earlier today I spotted that the 'Back' link shown on 'What
was their date of birth?' was hard-coded, despite the page
having several pages that route to it. This changes the link
to be dynamic using the approach in place for other such pages
in the app (i.e. we set it on the session).